% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODWGviz.r
\docType{package}
\name{ODWGviz-package}
\alias{ODWGviz}
\alias{ODWGviz-package}
\title{ODWGviz}
\description{
Package for testing and visualizing outlier detection algorithms.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
